requirements.txt
playwright>=1.40
pyyaml>=6.0
google-genai>=1.0
python-dotenv>=1.0
